a2a-sec deploy/requirements.txt
a2a-sdk==0.3.9
absolufy-imports==0.3.1
alembic==1.17.0
Authlib==1.6.1
cachetools==5.5.2
certifi==2025.7.14
cffi==1.17.1
cloudpickle==3.1.1
cryptography==45.0.5
docstring_parser==0.17.0
fastapi==0.116.1
google-adk==1.17.0
google-api-python-client==2.176.0
google-auth-httplib2==0.2.0
google-cloud-aiplatform==1.120.0
google-cloud-appengine-logging==1.6.2
google-cloud-aiplatform[agent_engines, adk]>=1.112.0
google-cloud-audit-log==0.3.2
google-cloud-bigquery==3.35.0
google-cloud-bigtable==2.33.0
google-cloud-discoveryengine==0.13.12
google-cloud-logging==3.12.1
google-cloud-monitoring==2.27.2
google-cloud-resource-manager==1.14.2
google-cloud-secret-manager==2.24.0
google-cloud-spanner==3.58.0
google-cloud-speech==2.33.0
google-cloud-storage==2.19.0
google-cloud-trace==1.16.2
google-genai==1.43.0
graphviz==0.21
grpc-interceptor==0.15.4
httpx==0.28.1
httpx-sse==0.4.1
Mako==1.3.10
mcp==1.12.0
numpy==2.3.1
opentelemetry-exporter-gcp-logging==1.9.0a0
opentelemetry-exporter-gcp-monitoring==1.9.0a0
opentelemetry-exporter-gcp-trace==1.9.0
opentelemetry-exporter-otlp-proto-http==1.37.0
opentelemetry-resourcedetector-gcp==1.9.0a0
pydantic==2.11.7
pydantic-settings==2.10.1
python-dotenv==1.1.1
python-multipart==0.0.20
PyYAML==6.0.2
requests==2.32.4
shapely==2.1.1
SQLAlchemy==2.0.41
sqlalchemy-spanner==1.17.0
sqlparse==0.5.3
sse-starlette==2.4.1
tenacity==8.5.0
tzlocal==5.3.1
urllib3==2.5.0
uvicorn==0.35.0
watchdog==6.0.0
websockets==15.0.1